agent-creator/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentcreator"
version = "1.0.0"
description = "Meta-agent for generating custom AI agents from SOPs"
requires-python = ">=3.12"

dependencies = [
    "langgraph>=0.5.4",
    "langchain-core>=0.3.72",
    "langchain-aws>=0.2.29",
    "langchain-anthropic>=0.3.17",
    "langchain-community>=0.3.31",
    "pydantic>=2.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "anthropic>=0.57.1",
    "python-multipart>=0.0.6",
    "boto3>=1.39.0",
    "bedrock-agentcore>=0.1.1",
    "opentelemetry-instrumentation-langchain",
    "aws-opentelemetry-distro~=0.10.1",
    "dspy-ai>=2.5.0",
    "beautifulsoup4>=4.14.2",
    "lxml>=4.9.0",
    "mcp>=1.16.0",
    "mlflow>=3.4.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["agentcreator*"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
